Split the file between times function and time operator
